+++ inst/yupana/pkgs.R
@@ -4,7 +4,7 @@
 #> open https://flavjack.github.io/inti/
 #> open https://flavjack.shinyapps.io/yupanapro/
 #> author .: Flavio Lozano-Isla (lozanoisla.com)
-#> date .: 2021-04-29
+#> date .: 2021-06-04
 # -------------------------------------------------------------------------
 
 # -------------------------------------------------------------------------
@@ -21,4 +21,5 @@ library(cowplot)
 library(corrplot)
 library(tidyverse)
 library(FactoMineR)
+library(ggpmisc) # add equation in regression